+++ BaseLib/ListTemplateMinimal.h
@@ -6,7 +6,7 @@ template <class M>
 class CListTemplateMinimal
 {
 protected:
-	int		miNumElements;
+	size_t	miNumElements;
 	int		miRowLength;  //The number of characters in the string including all terminating zeros.
 	size_t	miTotalSize;
 
@@ -22,8 +22,8 @@ public:
 	void	Init(int iNumFields, int iRowLength);
 	void	Kill(void);
 
-	M*		Get(int iIndex);
-	M*		Get(int iIndex, size_t* piSize);
+	M*		Get(size_t iIndex);
+	M*		Get(size_t iIndex, size_t* piSize);
 	M*		Add(M* pv, size_t iSize);
 	int		NumElements(void);
 	int		AllocatedElements(void);
@@ -76,7 +76,7 @@ void CListTemplateMinimal<M>::Kill(void)
 //																		//
 //////////////////////////////////////////////////////////////////////////
 template<class M>
-M* CListTemplateMinimal<M>::Get(int iIndex)
+M* CListTemplateMinimal<M>::Get(size_t iIndex)
 {
 	return (M*)RemapSinglePointer(this, maiFieldOffsets[iIndex]);
 }
@@ -87,7 +87,7 @@ M* CListTemplateMinimal<M>::Get(int iIndex)
 //																		//
 //////////////////////////////////////////////////////////////////////////
 template<class M>
-M* CListTemplateMinimal<M>::Get(int iIndex, size_t* piSize)
+M* CListTemplateMinimal<M>::Get(size_t iIndex, size_t* piSize)
 {
 	if (iIndex < miNumElements - 1)
 	{
